+++ previous-versions/sync-for-expenses-version-1/RELEASES.md
@@ -19,3 +19,13 @@ Based on:
 - [python v0.2.1] previous-versions/sync-for-expenses-version-1
 ### Releases
 - [PyPI v0.2.1] https://pypi.org/project/codat-sync-for-expenses-version-1/0.2.1 - previous-versions/sync-for-expenses-version-1
+
+## 2023-09-29 00:09:31
+### Changes
+Based on:
+- OpenAPI Doc prealpha https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Sync-Expenses-v1.yaml
+- Speakeasy CLI 1.91.3 (2.139.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v0.3.0] previous-versions/sync-for-expenses-version-1
+### Releases
+- [PyPI v0.3.0] https://pypi.org/project/codat-sync-for-expenses-version-1/0.3.0 - previous-versions/sync-for-expenses-version-1